* Change exceptionCode after merge pr * Fix Docker for tests * Some type fix * Fix types: max_execution_time & setConnectTimeOut

+++ src/Client.php
@@ -191,7 +191,7 @@ class Client
     /**
      * @return float
      */
-    public function getTimeout(): float
+    public function getTimeout(): int
     {
         return $this->settings()->getTimeOut();
     }
